+++ CHANGES.md
@@ -1,4 +1,7 @@
 ### Unreleased
+- Add networkInterfaceById and fileSystemById to GraphQL-API
+- Add one, five and fifteen LoadAverages to GraphQL-API.
+- Add monitors for load averages
 
 ### 0.18.3
 - Fixed: querying network interfaces on Windows takes too long